AI/video_detection/requirements.txt
flask==3.0.2
ultralytics==8.1.2
opencv-python==4.9.0.80
pillow==10.2.0
imagehash==4.3.1 